Add Dialogue Box When Created & Deleted

+++ PixelPlay Store/wwwroot/js/game-delete.js
@@ -2,15 +2,46 @@
     $('.js-delete').on('click', function () {
         var btn = $(this);
 
-        $.ajax({
-            url: `/Games/Delete/${btn.data('id')}`,
-            method: 'DELETE',
-            success: function () {
-                alert('success')
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+                confirmButton: "btn btn-danger",
+                cancelButton: "btn btn-dark mx-2 border-light"
             },
-            error: function () {
-                alert('error')
-            }
+            buttonsStyling: false
         });
+        swalWithBootstrapButtons.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonText: "Yes, delete it!",
+            cancelButtonText: "No, cancel!",
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    url: `/Games/Delete/${btn.data('id')}`,
+                    method: 'DELETE',
+                    success: function () {
+                        swalWithBootstrapButtons.fire({
+                            title: "Deleted!",
+                            text: "Game has been deleted.",
+                            icon: "success"
+                        }).then((result) => {
+                            if (result.isConfirmed) {
+                                window.location.href = '/Games'; 
+                            }
+                        });
+                    },
+                    error: function () {
+                        swalWithBootstrapButtons.fire({
+                            title: "Oops!",
+                            text: "Something went wrong.",
+                            icon: "error"
+                        });
+                    }
+                });
+            };               
+        });       
     });
 });
